feat(signers-page): wrapped sus components in error+sus boundaries
