+++ src/app/_components/ErrorBoundary.tsx
@@ -3,7 +3,6 @@ import { ComponentType, Fragment, PropsWithChildren, ReactNode } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 
 import { ExplorerError } from '../../common/types/Error';
-import { Box } from '../../ui/Box';
 import { ErrorBox } from './ErrorBox';
 
 interface ErrorBoundaryProps<WrapperProps extends PropsWithChildren<Record<string, unknown>>> {
