+++ src/app/signers/SignersHeader.tsx
@@ -1,224 +1,21 @@
-import { ArrowDownRight, ArrowRight, ArrowUpRight } from '@phosphor-icons/react';
-import { UseQueryResult, useQueries, useQueryClient } from '@tanstack/react-query';
-import pluralize from 'pluralize';
-import { ReactNode, useMemo, useState } from 'react';
+import { ArrowRight } from '@phosphor-icons/react';
+import { ReactNode, useState } from 'react';
 
 import { Card } from '../../common/components/Card';
-import { ApiResponseWithResultsOffset } from '../../common/types/api';
 import { TokenPrice } from '../../common/types/tokenPrice';
-import { numberToString } from '../../common/utils/utils';
 import { Box } from '../../ui/Box';
 import { Flex } from '../../ui/Flex';
 import { Icon } from '../../ui/Icon';
 import { Link } from '../../ui/Link';
 import { Stack } from '../../ui/Stack';
 import { Text } from '../../ui/Text';
-import BitcoinIcon from '../../ui/icons/BitcoinIcon';
-import { ExplorerErrorBoundary } from '../_components/ErrorBoundary';
-import { useSuspenseCurrentStackingCycle } from '../_components/Stats/CurrentStackingCycle/useCurrentStackingCycle';
-import { useSuspenseNextStackingCycle } from '../_components/Stats/NextStackingCycle/useNextStackingCycle';
+import { AddressesStackingCard } from './AddressesStackingCard';
 import { CurrentCycleCard } from './CurrentCycle';
+import { NextCycleCard } from './NextCycleCard';
 import { SignersDistribution } from './SignerDistribution';
 import { SignerDistributionHeader } from './SignerDistributionHeader';
-import { SignersStackersData, useGetStackersBySignerQuery } from './data/UseSignerAddresses';
-import { useStxSupply } from './data/usStxSupply';
-import { useSuspensePoxSigners } from './data/useSigners';
-
-function StatCardBase({
-  statTitle,
-  statValue,
-  moreInfo,
-}: {
-  statTitle: string;
-  statValue: string;
-  moreInfo: string | ReactNode;
-}) {
-  return (
-    <Card padding={6} height="100%">
-      <Stack gap={3}>
-        <Text fontSize="xs" fontWeight="medium" whiteSpace="nowrap">
-          {statTitle}
-        </Text>
-        <Text fontSize="xl" fontWeight="medium" whiteSpace="nowrap" display="inline-block" mr={1}>
-          {statValue}
-        </Text>
-        {typeof moreInfo === 'string' ? (
-          <Text fontSize="xs" fontWeight="medium" color="textSubdued" lineHeight={4}>
-            {moreInfo}
-          </Text>
-        ) : (
-          moreInfo
-        )}
-      </Stack>
-    </Card>
-  );
-}
-
-function StxStackedCard({ tokenPrice }: { tokenPrice: TokenPrice }) {
-  const { stackedSupply } = useStxSupply();
-  const stackedStxUsdValue =
-    tokenPrice.stxPrice != null ? tokenPrice.stxPrice * stackedSupply : undefined;
-  const stackedStxUsdValueFormatted = stackedStxUsdValue
-    ? `$${Math.round(stackedStxUsdValue).toLocaleString()}`
-    : undefined;
-  const stackedStxBtcValue =
-    stackedStxUsdValue && tokenPrice.btcPrice != null && tokenPrice.btcPrice !== 0
-      ? stackedStxUsdValue / tokenPrice.btcPrice
-      : undefined;
-  const stackedStxBtcValueFormatted = stackedStxBtcValue
-    ? `${stackedStxBtcValue.toFixed(1)} BTC`
-    : undefined;
-  const moreInfo =
-    stackedStxUsdValueFormatted && stackedStxBtcValueFormatted
-      ? `${stackedStxUsdValueFormatted} / ${stackedStxBtcValueFormatted}`
-      : undefined;
-
-  return (
-    <StatCardBase
-      statTitle="STX stacked"
-      statValue={numberToString(stackedSupply)}
-      moreInfo={
-        moreInfo ? (
-          <Text
-            fontSize="xs"
-            fontWeight="medium"
-            color="textSubdued"
-            whiteSpace="nowrap"
-            lineHeight={4}
-          >
-            {moreInfo}
-          </Text>
-        ) : null
-      }
-    />
-  );
-}
-
-function TotalStackedCard() {
-  const { stackedSupply, circulatingSupply } = useStxSupply();
-
-  const stxStackedPercentageFormatted = `${((stackedSupply / circulatingSupply) * 100).toFixed(
-    1
-  )}%`;
-
-  return (
-    <StatCardBase
-      statTitle="Total stacked"
-      statValue={stxStackedPercentageFormatted}
-      moreInfo={`/ ${numberToString(circulatingSupply)} circulating supply`}
-    />
-  );
-}
-
-function AddressesStackingCard() {
-  const { currentCycleId } = useSuspenseCurrentStackingCycle();
-  const previousCycleId = useMemo(() => currentCycleId - 1, [currentCycleId]);
-
-  const {
-    data: { results: currentCycleSigners },
-  } = useSuspensePoxSigners(currentCycleId);
-  const {
-    data: { results: previousCycleSigners },
-  } = useSuspensePoxSigners(previousCycleId);
-
-  const queryClient = useQueryClient();
-  const getQuery = useGetStackersBySignerQuery();
-  const currentCycleSignersStackersQueries = useMemo(() => {
-    return {
-      queries: currentCycleSigners.map(signer => {
-        return getQuery(currentCycleId, signer.signing_key);
-      }),
-      combine: (
-        response: UseQueryResult<ApiResponseWithResultsOffset<SignersStackersData>, Error>[]
-      ) => response.map(r => r.data?.results ?? []),
-    };
-  }, [currentCycleSigners, getQuery, currentCycleId]);
-  const previousCycleSignersStackersQueries = useMemo(() => {
-    return {
-      queries: previousCycleSigners
-        ? previousCycleSigners.map(signer => {
-            return getQuery(previousCycleId, signer.signing_key);
-          })
-        : [],
-      combine: (
-        response: UseQueryResult<ApiResponseWithResultsOffset<SignersStackersData>, Error>[]
-      ) => response.map(r => r.data?.results ?? []),
-    };
-  }, [previousCycleSigners, getQuery, previousCycleId]);
-
-  const currentCycleSignersStackers = useQueries(currentCycleSignersStackersQueries, queryClient);
-  const previousCycleSignersStackers = useQueries(previousCycleSignersStackersQueries, queryClient);
-
-  const numCurrentCycleStackers = currentCycleSignersStackers.length;
-  const numPreviousCycleStackers = previousCycleSignersStackers.length;
-
-  const rate = numPreviousCycleStackers
-    ? (numCurrentCycleStackers - numPreviousCycleStackers) / numPreviousCycleStackers
-    : undefined;
-
-  const moreInfo = rate ? (
-    <Text lineHeight={4} fontSize="xs" fontWeight="medium" color="textSubdued">
-      <Text display="inline" whiteSpace="nowrap">
-        <Icon
-          as={rate > 0 ? ArrowUpRight : ArrowDownRight}
-          size={3}
-          color={rate > 0 ? 'green.600' : 'red.600'}
-        />
-        &nbsp;{`${rate * 100}%`}&nbsp;
-      </Text>
-      <Text display="inline">{`${rate > 0 ? 'more' : 'less'} than previous cycle`}</Text>
-    </Text>
-  ) : null;
-
-  return (
-    <StatCardBase
-      statTitle="Addresses stacking"
-      statValue={numCurrentCycleStackers.toString()}
-      moreInfo={moreInfo}
-    />
-  );
-}
-
-function NextCycleCard() {
-  const {
-    currentRewardCycleId,
-    nextCycleBurnBlockHeightStart,
-    displayPreparePhaseInfo,
-    approximateDaysTilNextCyclePreparePhase,
-    approximateDaysTilNextCycleRewardPhase,
-  } = useSuspenseNextStackingCycle();
-
-  const moreInfo = (
-    <Text lineHeight={4} fontSize="xs" fontWeight="medium" color="textSubdued">
-      <Text display="inline">
-        {`Starts in ~${
-          displayPreparePhaseInfo
-            ? approximateDaysTilNextCyclePreparePhase
-            : approximateDaysTilNextCycleRewardPhase
-        } ${pluralize(
-          'day',
-          displayPreparePhaseInfo
-            ? approximateDaysTilNextCyclePreparePhase
-            : approximateDaysTilNextCycleRewardPhase
-        )}
-        at`}
-        &nbsp;
-      </Text>
-      <Text display="inline" whiteSpace="nowrap">
-        <Icon as={BitcoinIcon} size={3} />
-        &nbsp;{`#${nextCycleBurnBlockHeightStart}`}
-      </Text>
-    </Text>
-  );
-
-  return (
-    <StatCardBase
-      statTitle="Next cycle"
-      statValue={currentRewardCycleId.toString()}
-      moreInfo={moreInfo}
-    />
-  );
-}
+import { StxStackedCard } from './StxStackedCard';
+import { TotalStackedCard } from './TotalStackedCard';
 
 export function SignersHeaderLayout({
   stackingTitle,
@@ -335,13 +132,5 @@ export function SignersHeader({ tokenPrice }: { tokenPrice: TokenPrice }) {
         </Flex>
       }
     />
-  );
-}
-
-export function SignersHeaderWithErrorBoundary({ tokenPrice }: { tokenPrice: TokenPrice }) {
-  return (
-    <ExplorerErrorBoundary renderContent={() => null}>
-      <SignersHeader tokenPrice={tokenPrice} />
-    </ExplorerErrorBoundary>
   );
 }
