+++ src/app/signers/StatsCardBase.tsx
@@ -1,0 +1,35 @@
+import { ReactNode } from 'react';
+
+import { Card } from '../../common/components/Card';
+import { Stack } from '../../ui/Stack';
+import { Text } from '../../ui/Text';
+
+export function StatCardBase({
+  statTitle,
+  statValue,
+  moreInfo,
+}: {
+  statTitle: string;
+  statValue: string;
+  moreInfo: string | ReactNode;
+}) {
+  return (
+    <Card padding={6} height="100%">
+      <Stack gap={3}>
+        <Text fontSize="xs" fontWeight="medium" whiteSpace="nowrap">
+          {statTitle}
+        </Text>
+        <Text fontSize="xl" fontWeight="medium" whiteSpace="nowrap" display="inline-block" mr={1}>
+          {statValue}
+        </Text>
+        {typeof moreInfo === 'string' ? (
+          <Text fontSize="xs" fontWeight="medium" color="textSubdued" lineHeight={4}>
+            {moreInfo}
+          </Text>
+        ) : (
+          moreInfo
+        )}
+      </Stack>
+    </Card>
+  );
+}
